+++ src/main/java/absolutelyaya/ultracraft/block/mapping/ForceTravelBlockEntity.java
@@ -23,7 +23,7 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 {
 	List<? extends LivingEntity> lastContained = new ArrayList<>();
 	static List<String> attributes = new ArrayList<>();
-	boolean openRanking = true;
+	boolean openRanking = true, rankingTitleSuffix = true;
 	
 	public ForceTravelBlockEntity(BlockPos pos, BlockState state)
 	{
@@ -60,6 +60,8 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 	{
 		if(attribute.equals("openRanking"))
 			return String.valueOf(openRanking);
+		else if(attribute.equals("rankingTitleSuffix"))
+			return String.valueOf(rankingTitleSuffix);
 		return null;
 	}
 	
@@ -68,6 +70,8 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 	{
 		if(s.equals("openRanking"))
 			openRanking = Boolean.parseBoolean(value);
+		else if(s.equals("rankingTitleSuffix"))
+			rankingTitleSuffix = Boolean.parseBoolean(value);
 		super.setAttribute(s, value);
 	}
 	
@@ -84,6 +88,7 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 					PacketByteBuf buf = new PacketByteBuf(Unpooled.buffer());
 					buf.writeBoolean(true);
 					buf.writeBoolean(openRanking);
+					buf.writeBoolean(rankingTitleSuffix);
 					ServerPlayNetworking.send(player, PacketRegistry.TRAVEL_SCREEN_PACKET_ID, buf);
 					UltraComponents.LEVEL_STATS.get(player).onFinishLevel();
 				}
@@ -98,6 +103,8 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 		super.readNbt(nbt);
 		if(nbt.contains("openRanking", NbtElement.BYTE_TYPE))
 			openRanking = nbt.getBoolean("openRanking");
+		if(nbt.contains("rankingTitleSuffix", NbtElement.BYTE_TYPE))
+			rankingTitleSuffix = nbt.getBoolean("rankingTitleSuffix");
 	}
 	
 	@Override
@@ -105,6 +112,7 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 	{
 		super.writeNbt(nbt);
 		nbt.putBoolean("openRanking", openRanking);
+		nbt.putBoolean("rankingTitleSuffix", rankingTitleSuffix);
 	}
 	
 	@Override
@@ -115,5 +123,6 @@ public class ForceTravelBlockEntity extends AbstractTriggerBlockEntity
 	
 	static {
 		attributes.add("openRanking");
+		attributes.add("rankingTitleSuffix");
 	}
 }